+++ tests/testthat/test-cre.R
@@ -26,7 +26,6 @@ test_that("cre Runs Correctly", {
                       t_ext = 0.025,
                       t_corr = 1,
                       t_pvalue = 0.05,
-                      replace = FALSE,
                       stability_selection = TRUE,
                       cutoff = 0.6,
                       pfer = 1,
@@ -70,10 +69,6 @@ test_that("cre Runs Correctly", {
   expect_error(cre(y, z, X, method_params, hyper_params))
 
   hyper_params[["node_size"]] <- 5
-  hyper_params[["replace"]] <- "test"
-  expect_error(cre(y, z, X, method_params, hyper_params))
-
-  hyper_params[["replace"]] <- TRUE
   hyper_params[["max_nodes"]] <- "test"
   expect_error(cre(y, z, X, method_params, hyper_params))