+++ tests/testthat/test-filter_extreme_rules.R
@@ -15,7 +15,6 @@ test_that("Extreme Rules Discarded Correctly", {
   node_size <- 20
   max_nodes <- 5
   max_depth <- 15
-  replace <- FALSE
   t_decay <- 0.025
   t_ext <- 0.1
 
@@ -33,7 +32,7 @@ test_that("Extreme Rules Discarded Correctly", {
 
   # Step 3: Generate rules list
   initial_rules <- generate_rules(X, ite, ntrees_rf, ntrees_gbm, node_size,
-                                  max_nodes, max_depth, replace)
+                                  max_nodes, max_depth)
 
   rules_list <- filter_irrelevant_rules(initial_rules, X, ite, t_decay)
   rules_matrix <- generate_rules_matrix(X, rules_list)