+++ tests/testthat/test-generate_rules_matrix.R
@@ -15,7 +15,6 @@ test_that("Rules Extracted Correctly", {
   node_size <- 20
   max_nodes <- 5
   max_depth <- 15
-  replace <- FALSE
   t_decay <- 0.025
 
   # Step 1: Split data
@@ -32,7 +31,7 @@ test_that("Rules Extracted Correctly", {
 
   # Step 3: Generate rules list
   initial_rules <- generate_rules(X, ite, ntrees_rf, ntrees_gbm, node_size,
-                                  max_nodes, max_depth, replace)
+                                  max_nodes, max_depth)
 
   rules <- filter_irrelevant_rules(initial_rules, X, ite, t_decay)